test some features of jsonarray
